requirements.txt
deepface
opencv-python
requests
tensorflow-cpu
tf-keras
python-dotenv
